pyproject.toml
[project]
name = "workbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "curl-cffi==0.15.0b4",
    "discord-py>=2.7.1",
    "nodriver>=0.48.1",
    "pickles>=0.1.1",
    "playwright==1.58.0",
    "playwright-stealth>=2.0.2",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
